This PR does the following changes:
- refactors code and better places EIP-55 checksum on the `Address` type
- makes all API `Address` occurrences serialize in the EIP-55 string format

Closes #6725
