+++ common/primitive-types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hopr-primitive-types"
-version = "0.5.0"
+version = "0.6.0"
 authors = ["HOPR Association <tech@hoprnet.org>"]
 edition = "2021"
 description = "Generic types used through the entire code base"
@@ -20,6 +20,7 @@ rlp = "0.5.2"
 regex = { workspace = true }
 strum = { workspace = true }
 serde = { workspace = true }
+sha3 = { workspace = true }
 thiserror = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
