+++ common/primitive-types/src/primitives.rs
@@ -2,6 +2,7 @@ use chrono::serde::ts_seconds_option;
 use chrono::{DateTime, Utc};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
+use sha3::Digest;
 use std::cmp::Ordering;
 use std::fmt::{Debug, Display, Formatter};
 use std::ops::{Add, Mul, Sub};
@@ -48,6 +49,27 @@ impl Address {
     /// Checks if the address is all zeroes.
     pub fn is_zero(&self) -> bool {
         self.0.iter().all(|e| 0_u8.eq(e))
+    }
+
+    /// Turns the address into a checksum-ed address string
+    /// according to https://eips.ethereum.org/EIPS/eip-55>
+    pub fn to_checksum(&self) -> String {
+        let address_hex = hex::encode(self.0);
+
+        let hash = sha3::Keccak256::digest(address_hex.as_bytes());
+
+        let mut ret = String::with_capacity(Self::SIZE * 2 + 2);
+        ret.push_str("0x");
+
+        for (i, c) in address_hex.chars().enumerate() {
+            let nibble = hash[i / 2] >> (((i + 1) % 2) * 4) & 0xf;
+            if nibble >= 8 {
+                ret.push(c.to_ascii_uppercase());
+            } else {
+                ret.push(c);
+            }
+        }
+        ret
     }
 }
 
@@ -638,5 +660,75 @@ mod tests {
         let expected_le = hex!("081A99BE1C000000000000000000000000000000000000000000000000000000");
         assert_eq!(expected_le, le_bytes);
         assert_eq!(U256::from_le_bytes(expected_le), num);
+    }
+
+    #[test]
+    fn test_address_to_checksum_all_caps() -> anyhow::Result<()> {
+        let addr_1 = Address::from_str("52908400098527886e0f7030069857d2e4169ee7")?;
+        let value_1 = addr_1.to_checksum();
+        let addr_2 = Address::from_str("8617e340b3d01fa5f11f306f4090fd50e238070d")?;
+        let value_2 = addr_2.to_checksum();
+
+        assert_eq!(
+            value_1, "0x52908400098527886E0F7030069857D2E4169EE7",
+            "checksumed address does not match"
+        );
+        assert_eq!(
+            value_2, "0x8617E340B3D01FA5F11F306F4090FD50E238070D",
+            "checksumed address does not match"
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_address_to_checksum_all_lower() -> anyhow::Result<()> {
+        let addr_1 = Address::from_str("de709f2102306220921060314715629080e2fb77")?;
+        let value_1 = addr_1.to_checksum();
+        let addr_2 = Address::from_str("27b1fdb04752bbc536007a920d24acb045561c26")?;
+        let value_2 = addr_2.to_checksum();
+
+        assert_eq!(
+            value_1, "0xde709f2102306220921060314715629080e2fb77",
+            "checksumed address does not match"
+        );
+        assert_eq!(
+            value_2, "0x27b1fdb04752bbc536007a920d24acb045561c26",
+            "checksumed address does not match"
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_address_to_checksum_all_normal() -> anyhow::Result<()> {
+        let addr_1 = Address::from_str("5aaeb6053f3e94c9b9a09f33669435e7ef1beaed")?;
+        let addr_2 = Address::from_str("fb6916095ca1df60bb79ce92ce3ea74c37c5d359")?;
+        let addr_3 = Address::from_str("dbf03b407c01e7cd3cbea99509d93f8dddc8c6fb")?;
+        let addr_4 = Address::from_str("d1220a0cf47c7b9be7a2e6ba89f429762e7b9adb")?;
+
+        let value_1 = addr_1.to_checksum();
+        let value_2 = addr_2.to_checksum();
+        let value_3 = addr_3.to_checksum();
+        let value_4 = addr_4.to_checksum();
+
+        assert_eq!(
+            value_1, "0x5aAeb6053F3E94C9b9A09f33669435E7Ef1BeAed",
+            "checksumed address does not match"
+        );
+        assert_eq!(
+            value_2, "0xfB6916095ca1df60bB79Ce92cE3Ea74c37c5d359",
+            "checksumed address does not match"
+        );
+        assert_eq!(
+            value_3, "0xdbF03B407c01E7cD3CBea99509d93f8DDDC8C6FB",
+            "checksumed address does not match"
+        );
+        assert_eq!(
+            value_4, "0xD1220A0cf47c7B9Be7A2E6BA89F429762e7b9aDb",
+            "checksumed address does not match"
+        );
+
+        Ok(())
     }
 }
