+++ crypto/types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hopr-crypto-types"
-version = "0.1.0"
+version = "0.1.1"
 description = "Implements basic cryptograpic primitives and cryptography related types"
 edition = "2021"
 
@@ -27,7 +27,7 @@ libp2p-identity = { workspace = true }
 primitive-types = { workspace = true }
 serde = { workspace = true }
 sha2 = "0.10.8"
-sha3 = "0.10.8"
+sha3 = { workspace = true }
 subtle = "2.5"
 thiserror = { workspace = true }
 tracing = { workspace = true }
